requirements.txt
streamlit>=1.41.0
yfinance>=0.2.50
p123api>=2.1.0
python-dotenv>=1.0.1
requests>=2.28.0
